+++ PanoramicData.OData.Client/ODataClient.cs
@@ -21,6 +21,7 @@ public partial class ODataClient : IDisposable
 		DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
 		Converters =
 		{
+			new ODataTypeAnnotationConverter(), // Must be first to handle @odata.type annotation
 			new JsonStringEnumConverter(),
 			new ODataDateTimeConverter(),
 			new ODataNullableDateTimeConverter()
@@ -35,6 +36,7 @@ public partial class ODataClient : IDisposable
 		DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
 		Converters =
 		{
+			new ODataTypeAnnotationConverter(), // Must be first to handle @odata.type annotation
 			new JsonStringEnumConverter(),
 			new ODataDateTimeConverter(),
 			new ODataNullableDateTimeConverter()